requirements.txt
# Common dependances
cryptography
mysql-connector-python

# WebAPI dependances
fastapi
python-multipart
uvicorn

# SMTP server dependances
aiosmtpd
